pyproject.toml
[project]
name = "fast-api-demo-104"
version = "0.1.0"
description = "This project is a demo project for 104 interview, which mainly targeting FastAPI exploration."
authors = [
    {name = "Huang Chuan",email = "river1440@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "nanoid (>=2.0.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
pre-commit = "^4.2.0"


[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
factory-boy = "^3.3.3"
faker = "^37.1.0"
pytest-asyncio = "^0.26.0"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "."
